skip test 02911,02908 and 01158

+++ docker/test/stateless/run.sh
@@ -210,10 +210,18 @@ function run_tests()
     try_run_with_retry 10 clickhouse-client -q "insert into system.zookeeper (name, path, value) values ('api_version', '/keeper/api_version/', 2)"
 
     set +e
-    clickhouse-test --testname --shard --zookeeper --check-zookeeper-session --hung-check --print-time \
+    if [[ -f "setup_fdb.sh" ]] && [[ "$(uname -m)" == "x86_64" ]]; then
+        clickhouse-test --testname --shard --zookeeper --check-zookeeper-session --hung-check --print-time \
+        --test-runs "$NUM_TRIES" "${ADDITIONAL_OPTIONS[@]}" \
+        --skip 02911_backup_restore_keeper_map 02908_many_requests_to_system_replicas 01158_zookeeper_log_long 2>&1 \
+        | ts '%Y-%m-%d %H:%M:%S' \
+        | tee -a test_output/test_result.txt
+    else
+        clickhouse-test --testname --shard --zookeeper --check-zookeeper-session --hung-check --print-time \
         --test-runs "$NUM_TRIES" "${ADDITIONAL_OPTIONS[@]}" 2>&1 \
-    | ts '%Y-%m-%d %H:%M:%S' \
-    | tee -a test_output/test_result.txt
+        | ts '%Y-%m-%d %H:%M:%S' \
+        | tee -a test_output/test_result.txt
+    fi
     set -e
 }
